drm/lima: fix submit wait error in arm64
This error is due to assigned long return value
to int which may cause overflow in 64bit arch.

+++ drivers/gpu/drm/lima/lima_gem.c
@@ -530,7 +530,7 @@ int lima_gem_wait(struct drm_file *file, u32 handle, u32 op, u64 timeout_ns)
 	bool write = op & LIMA_GEM_WAIT_WRITE;
 	struct drm_gem_object *obj;
 	struct lima_bo *bo;
-	int ret;
+	signed long ret;
 	unsigned long timeout;
 
 	obj = drm_gem_object_lookup(file, handle);
@@ -540,6 +540,10 @@ int lima_gem_wait(struct drm_file *file, u32 handle, u32 op, u64 timeout_ns)
 	bo = to_lima_bo(obj);
 
 	timeout = timeout_ns ? lima_timeout_to_jiffies(timeout_ns) : 0;
+
+	/* must use long for result check because in 64bit arch int
+	 * will overflow if timeout is too large and get <0 result
+	 */
 	ret = reservation_object_wait_timeout_rcu(bo->resv, write, true, timeout);
 	if (ret == 0)
 		ret = timeout ? -ETIMEDOUT : -EBUSY;
